fix Cannot find module when import ResourceType in server from public folder (#1705)
* fix Cannot find module


* fix prettier error


* fix failed tests


---------

(cherry picked from commit 09b2f59453b4b3d356ba7627d1e1af663fb075bd)
