+++ public/apps/configuration/panels/role-edit/cluster-permission-panel.tsx
@@ -15,7 +15,8 @@
 
 import React, { Dispatch, SetStateAction } from 'react';
 import { EuiForm, EuiFlexGroup, EuiFlexItem, EuiComboBox } from '@elastic/eui';
-import { ComboBoxOptions, ResourceType } from '../../types';
+import { ComboBoxOptions } from '../../types';
+import { ResourceType } from '../../../../../common';
 import { PanelWithHeader } from '../../utils/panel-with-header';
 import { FormRow } from '../../utils/form-row';
 import { LIMIT_WIDTH_INPUT_CLASS } from '../../constants';
