+++ public/apps/configuration/panels/role-edit/index-permission-panel.tsx
@@ -25,7 +25,8 @@ import {
 } from '@elastic/eui';
 import React, { Dispatch, Fragment, SetStateAction } from 'react';
 import { isEmpty } from 'lodash';
-import { RoleIndexPermission, ResourceType } from '../../types';
+import { RoleIndexPermission } from '../../types';
+import { ResourceType } from '../../../../../common';
 import {
   appendElementToArray,
   removeElementFromArray,
@@ -141,9 +142,9 @@ export function IndexPermissionRow(props: {
   return (
     <FormRow
       headerText="Index permissions"
-      headerSubText="You can specify permissions using both action groups or single permissions. 
+      headerSubText="You can specify permissions using both action groups or single permissions.
         A permission group is a list of single permissions.
-        You can often achieve your desired security posture using some combination of the default permission groups. 
+        You can often achieve your desired security posture using some combination of the default permission groups.
         You can also create your own reusable permission groups."
     >
       <EuiFlexGroup>
