+++ public/apps/configuration/panels/role-edit/role-edit.tsx
@@ -46,7 +46,8 @@ import {
 } from './tenant-panel';
 import { RoleIndexPermissionStateClass, RoleTenantPermissionStateClass } from './types';
 import { buildHashUrl, buildUrl } from '../../utils/url-builder';
-import { ComboBoxOptions, ResourceType, Action, ActionGroupItem } from '../../types';
+import { ComboBoxOptions, Action, ActionGroupItem } from '../../types';
+import { ResourceType } from '../../../../../common';
 import {
   useToastState,
   createUnknownErrorToast,
