+++ public/apps/configuration/panels/role-mapping/role-edit-mapped-user.tsx
@@ -35,7 +35,8 @@ import { ExternalIdentityStateClass } from './types';
 import { ComboBoxOptions } from '../../types';
 import { stringToComboBoxOption, comboBoxOptionToString } from '../../utils/combo-box-utils';
 import { buildHashUrl, buildUrl } from '../../utils/url-builder';
-import { ResourceType, RoleMappingDetail, SubAction, Action } from '../../types';
+import { RoleMappingDetail, SubAction, Action } from '../../types';
+import { ResourceType } from '../../../../../common';
 import { fetchUserNameList } from '../../utils/internal-user-list-utils';
 import { updateRoleMapping, getRoleMappingData } from '../../utils/role-mapping-utils';
 import { createErrorToast, createUnknownErrorToast, useToastState } from '../../utils/toast-utils';
