+++ public/apps/configuration/panels/role-mapping/users-panel.tsx
@@ -19,7 +19,8 @@ import { ComboBoxOptions } from '../../types';
 import { PanelWithHeader } from '../../utils/panel-with-header';
 import { FormRow } from '../../utils/form-row';
 import { buildHashUrl } from '../../utils/url-builder';
-import { ResourceType, Action } from '../../types';
+import { Action } from '../../types';
+import { ResourceType } from '../../../../../common';
 import { ExternalLinkButton } from '../../utils/display-utils';
 import { DocLinks } from '../../constants';
 import { appendOptionToComboBoxHandler } from '../../utils/combo-box-utils';
