+++ public/apps/configuration/panels/role-view/cluster-permission-panel.tsx
@@ -17,7 +17,8 @@ import React from 'react';
 import { EuiEmptyPrompt, EuiButton } from '@elastic/eui';
 import { PanelWithHeader } from '../../utils/panel-with-header';
 import { PermissionTree } from '../permission-tree';
-import { ActionGroupItem, DataObject, ResourceType, Action } from '../../types';
+import { ActionGroupItem, DataObject, Action } from '../../types';
+import { ResourceType } from '../../../../../common';
 import { buildHashUrl } from '../../utils/url-builder';
 import { loadingSpinner } from '../../utils/loading-spinner-utils';
 import { DocLinks } from '../../constants';
