+++ public/apps/configuration/panels/role-view/role-view.tsx
@@ -39,7 +39,6 @@ import { difference } from 'lodash';
 import { BreadcrumbsPageDependencies } from '../../../types';
 import { buildHashUrl, buildUrl } from '../../utils/url-builder';
 import {
-  ResourceType,
   Action,
   SubAction,
   RoleMappingDetail,
@@ -48,6 +47,7 @@ import {
   RoleIndexPermissionView,
   RoleTenantPermissionView,
 } from '../../types';
+import { ResourceType } from '../../../../../common';
 import {
   getRoleMappingData,
   MappedUsersListing,
