+++ public/apps/configuration/panels/role-view/test/role-view.test.tsx
@@ -30,7 +30,8 @@ import { getRoleDetail } from '../../../utils/role-detail-utils';
 import { transformRoleIndexPermissions } from '../../../utils/index-permission-utils';
 import { useDeleteConfirmState } from '../../../utils/delete-confirm-modal-utils';
 import { requestDeleteRoles } from '../../../utils/role-list-utils';
-import { Action, ResourceType, SubAction } from '../../../types';
+import { Action, SubAction } from '../../../types';
+import { ResourceType } from '../../../../../../common';
 import { buildHashUrl } from '../../../utils/url-builder';
 import { createUnknownErrorToast } from '../../../utils/toast-utils';
