+++ public/apps/configuration/panels/role-view/test/tenants-panel.test.tsx
@@ -18,7 +18,8 @@ import { mount, shallow } from 'enzyme';
 import { TenantsPanel } from '../tenants-panel';
 import { EuiEmptyPrompt, EuiInMemoryTable, EuiTableFieldDataColumnType } from '@elastic/eui';
 import { buildHashUrl } from '../../../utils/url-builder';
-import { Action, ResourceType, RoleTenantPermissionDetail } from '../../../types';
+import { Action, RoleTenantPermissionDetail } from '../../../types';
+import { ResourceType } from '../../../../../../common';
 import { RoleViewTenantInvalidText } from '../../../constants';
 
 jest.mock('../../../utils/tenant-utils');
