+++ public/apps/configuration/panels/user-list.tsx
@@ -35,7 +35,8 @@ import React, { useState } from 'react';
 import { getAuthInfo } from '../../../utils/auth-info-utils';
 import { AppDependencies } from '../../types';
 import { API_ENDPOINT_INTERNALUSERS, DocLinks } from '../constants';
-import { Action, ResourceType } from '../types';
+import { Action } from '../types';
+import { ResourceType } from '../../../../common';
 import { EMPTY_FIELD_VALUE } from '../ui-constants';
 import { useContextMenuState } from '../utils/context-menu';
 import { useDeleteConfirmState } from '../utils/delete-confirm-modal-utils';
